refactor: move variables and function to separate files
