+++ index.php
@@ -1,70 +1,18 @@
 <?php
 
+/**
+ * @var string[] $categories_list Список категорий
+ * @var array<int,array{name: string, category: string, price: int, img: string} $lots_list Список лотов
+ * @var bool $is_auth Флаг авторизации
+ * @var string $user_name Имя пользователя
+ */
+
 declare(strict_types=1);
 
 require_once 'helpers.php';
-
-/**
- * Форматирует цену добавляя разделители для тысяч и символ рубля
- *
- * @param int $price Не отформатированная цена
- * @return string Отформатированная цена с символом рубля
- */
-function format_price(int $price): string
-{
-    $formatted_price = number_format($price, 0, '', ' ');
-    return "$formatted_price ₽";
-}
+require_once 'data.php';
 
 $page_title = 'Главная';
-$is_auth = rand(0, 1);
-$user_name = 'Артём';
-$categories_list = [
-    'Доски и лыжи',
-    'Крепления',
-    'Ботинки',
-    'Одежда',
-    'Инструменты',
-    'Разное',
-];
-$lots_list = [
-    [
-        'name' => '2014 Rossignol District Snowboard',
-        'category' => 'Доски и лыжи',
-        'price' => 10999,
-        'img' => '/img/lot-1.jpg',
-    ],
-    [
-        'name' => 'DC Ply Mens 2016/2017 Snowboard',
-        'category' => 'Доски и лыжи',
-        'price' => 159999,
-        'img' => '/img/lot-2.jpg',
-    ],
-    [
-        'name' => 'Крепления Union Contact Pro 2015 года размер L/XL',
-        'category' => 'Крепления',
-        'price' => 8000,
-        'img' => '/img/lot-3.jpg',
-    ],
-    [
-        'name' => 'Ботинки для сноуборда DC Mutiny Charocal',
-        'category' => 'Ботинки',
-        'price' => 10999,
-        'img' => '/img/lot-4.jpg',
-    ],
-    [
-        'name' => 'Куртка для сноуборда DC Mutiny Charocal',
-        'category' => 'Одежда',
-        'price' => 7500,
-        'img' => '/img/lot-5.jpg',
-    ],
-    [
-        'name' => 'Маска Oakley Canopy',
-        'category' => 'Разное',
-        'price' => 5400,
-        'img' => '/img/lot-6.jpg',
-    ],
-];
 
 $page_content = include_template(
     'main.php',
